refactor(search): inject ranking nowMs and extract UnifiedRanking
Thread SearchRequest.nowMs through unified search so semantic recency
uses an injected clock, not Date(). Lift rerank transforms into
package enum UnifiedRanking and pin product SearchRequest construction.
